adding tests for taskmetadata_manager
